changed to doubles

+++ src/main/java/Calculator.java
@@ -1,6 +1,7 @@
 public class Calculator {
 
-    public static int multiply(int a, int b){
+    public static double multiply(double a, double b){
         return a * b;
     }
+
 }
